move string, merkle map, merkle tree to provable-types
