+++ src/lib/provable-types/string.ts
@@ -1,12 +1,8 @@
-import { Bool, Field } from '../lib/core.js';
-import { Provable } from './provable.js';
-import { Poseidon } from './hash.js';
-import { Gadgets } from './gadgets/gadgets.js';
-import {
-  CircuitValue,
-  arrayProp,
-  prop,
-} from './provable-types/circuit-value.js';
+import { Bool, Field } from '../core.js';
+import { Provable } from '../provable.js';
+import { Poseidon } from '../hash.js';
+import { Gadgets } from '../gadgets/gadgets.js';
+import { CircuitValue, arrayProp, prop } from './circuit-value.js';
 
 export { Character, CircuitString };
